fix depreciations + change price type

+++ src/Entity/Expense.php
@@ -32,7 +32,7 @@ class Expense
     #[Groups(["expense:read", "expense:write"])]
     private $description;
 
-    #[ORM\Column(type: 'float')]
+    #[ORM\Column(type: 'integer')]
     #[Groups(["expense:read", "expense:write"])]
     private $price;
 
@@ -86,12 +86,12 @@ class Expense
         return $this;
     }
 
-    public function getPrice(): ?float
+    public function getPrice(): ?int
     {
         return $this->price;
     }
 
-    public function setPrice(float $price): self
+    public function setPrice(int $price): self
     {
         $this->price = $price;
